crawling-server/app/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
redis==5.0.1
celery==5.3.4
flower==2.0.1
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3
playwright==1.40.0
pydantic==2.5.0
pydantic-settings==2.1.0
prometheus-client==0.22.1
sentry-sdk==2.34.1
python-dotenv==1.0.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.13.0
